Time: 577 ms (5.18%), Space: 102.9 MB (6.91%) - LeetHub

+++ 2126-destroying-asteroids/2126-destroying-asteroids.cpp
@@ -3,7 +3,7 @@ public:
     bool asteroidsDestroyed(int mass, vector<int>& asteroids) {
         sort(asteroids.begin(),asteroids.end());
         
-        unsigned long int m = (unsigned long int)mass;
+        unsigned long int m = mass;
         
         for(int i=0;i<asteroids.size();i++){
             if(m<asteroids[i])
